Start moving dotnet relative source tests to dotnet integration tests.
